requirements.txt
eventlet
Flask
python-socketio
pyserial
